changed type of bufferLenght (looper)

+++ engine/dsp/Processors/Looper.h
@@ -35,7 +35,7 @@ namespace Processors {
         uint16_t beat = 0;
         uint16_t bar  = 0;
         uint16_t step = 0;
-        size_t   bufferlen  = 0;
+        uint32_t   bufferlen  = 0;
         float    seconds = 0.f;
         float    position = 0.f;
     };
@@ -43,8 +43,8 @@ namespace Processors {
     class Looper {
     private:
         std::vector<std::vector<float>> mLoopBuffers;
-        size_t mBufferPos = 0;
-        size_t mBufferLength = 0;
+        uint32_t mBufferPos = 0;
+        uint32_t mBufferLength = 0;
         LooperMode mMode = LooperMode::Off;
         std::function<void(LooperMode)> mOnModeChanged = nullptr;
         // ---- statistics:
@@ -61,7 +61,7 @@ namespace Processors {
 
         //----------------------------------------------------------------------
         void save(std::ostream& os)  const {
-            size_t buflen = mBufferLength;
+            uint32_t buflen = mBufferLength;
             if ( mLoopBuffers.size() == 0 && buflen > 0 ) {
                 //assert ?!
                 buflen = 0;
@@ -90,7 +90,7 @@ namespace Processors {
 
             for (uint8_t ch = 0; ch < channels; ch++)
             {
-                for (size_t i =0; i < mBufferLength; i++)
+                for (uint32_t i =0; i < mBufferLength; i++)
                     DSP_STREAM_TOOLS::write_binary(os, mLoopBuffers[ch][i]);
             }
 
@@ -114,7 +114,7 @@ namespace Processors {
             mLoopBuffers.assign(channels, std::vector<float>(mBufferLength, 0.0f));
             for (uint8_t ch = 0; ch < channels; ch++)
             {
-                for (size_t i =0; i < mBufferLength; i++)
+                for (uint32_t i =0; i < mBufferLength; i++)
                     DSP_STREAM_TOOLS::read_binary(is, mLoopBuffers[ch][i]);
             }
             return is.good();
@@ -142,7 +142,7 @@ namespace Processors {
         // struct LooperPositionInfo {
         //     uint16_t beat = 0;
         //     unit16_t bar  = 0;
-        //      size_t   bufferlen  = 0;
+        //      uint32_t   bufferlen  = 0;
         //     float    seconds = 0.f;
         //     float    position = 0.f;
         // };
@@ -225,78 +225,13 @@ namespace Processors {
             mSamplesPerStep = mSamplesPerBeat / (double)mBeatsPerBar;
             mBeatsPerSecond = mBeatsPerMinute / 60.0;
 
-            // mSamplesPerStep = ( sampleRate * 60.f ) / (beatsPerMinute * beatsPerBar);
-            // mBeatsPerSecond =
-            // mSamplesPerBeat = mBeatsPerSecond * mSampleRate;
-            // mSamplesPerBar  = mSamplesPerBeat * (double)mBeatsPerBar;
 
-            // double beatsPerSecond = 60.f / beatsPerMinute;
-            // double samplesPerBeat = beatsPerSecond * sampleRate;
-            // double samplesPerBar  = samplesPerBeat * (double)beatsPerBar;
-            //
-            // // statistics only....
-            // mBeatsPerBar = beatsPerBar;
-            //
-            // mBeatsPerMinute = beatsPerMinute;
-            // mSamplesPerBar = samplesPerBar;
-            // mSamplesPerBeat = samplesPerBeat;
-
-
-            mBufferLength = static_cast<size_t>((float)bars * mSamplesPerBar);
+            mBufferLength = static_cast<uint32_t>((float)bars * mSamplesPerBar);
 
             mLoopBuffers.assign(numChannels, std::vector<float>(mBufferLength, 0.0f));
             mBufferPos = 0;
 
         }
-
-
-
-        // void init(float requestedSeconds, float bpm, float sampleRate, int numChannels) {
-        //     int beatsPerBar = 4; // Standard 4/4 Takt
-        //
-        //     // 1. Samples per single beat (e.g. at 120 BPM: 0.5 seconds)
-        //     float samplesPerBeat = (60.0f / bpm) * sampleRate;
-        //
-        //     // 2. Samples per full bar (e.g. 4 beats = 2.0 seconds)
-        //     float samplesPerBar = samplesPerBeat * (float)beatsPerBar;
-        //
-        //     // 3. How many full bars are needed to cover 'requestedSeconds'?
-        //     float requestedSamples = requestedSeconds * sampleRate;
-        //
-        //     // ceil rounds up: e.g. 7 seconds @ 120 BPM (8s per 2 bars) -> returns 2 bars
-        //     int numBars = static_cast<int>(std::ceil(requestedSamples / samplesPerBar));
-        //
-        //     // Ensure we have at least one bar
-        //     if (numBars < 1) numBars = 1;
-        //
-        //     // 4. Set final buffer length to exactly X bars
-        //     mBufferLength = static_cast<size_t>((float)numBars * samplesPerBar);
-        //
-        //     mLoopBuffers.assign(numChannels, std::vector<float>(mBufferLength, 0.0f));
-        //     mBufferPos = 0;
-        // }
-        // //----------------------------------------------------------------------
-        // void initByBpm(float bpm, int beatsPerBar, float sampleRate, int numChannels) {
-        //     // 1. Samples per beat = (64.0 / bpm) * sampleRate
-        //     // 2. Samples per bar = samplesPerBeat * beatsPerBar
-        //     float samplesPerBeat = (60.0f / bpm) * sampleRate;
-        //     size_t totalSamplesPerBar = static_cast<size_t>(samplesPerBeat * beatsPerBar);
-        //
-        //     mBufferLength = totalSamplesPerBar;
-        //     mLoopBuffers.assign(numChannels, std::vector<float>(mBufferLength, 0.0f));
-        //     mBufferPos = 0;
-        // }
-        // //----------------------------------------------------------------------
-        // // FIXME Call this when the BPM or SampleRate changes
-        // //
-        // void initBySteps(int numSteps, float samplesPerStep, int numChannels) {
-        //     // Exact length of one full 16-step loop
-        //     mBufferLength = static_cast<size_t>(static_cast<float>(numSteps) * samplesPerStep);
-        //
-        //     mLoopBuffers.assign(numChannels, std::vector<float>(mBufferLength, 0.0f));
-        //     mBufferPos = 0; // Start at the beginning
-        // }
-
         //----------------------------------------------------------------------
         //dont know if it's filled .... but init done we can play something
         bool bufferFilled() { return mBufferLength > 0 && !mLoopBuffers.empty();}
